// x-pack/platform/packages/shared/ai-assistant/ai-assistant-connector-selector-action/connector_selector.translations.d.ts
/**
 * Translations for the `ConnectorSelector` component.
 */
export declare const translations: {
    selectableAriaLabel: string;
    addConnectorAriaLabel: string;
    addConnectorLabel: string;
    manageConnectorAriaLabel: string;
    preConfiguredConnectorLabel: string;
    customConnectorLabel: string;
    defaultConnectorLabel: string;
};